+++ bot.py
@@ -18,53 +18,66 @@
 
 import logging
 import os
+from datetime import datetime
 from math import ceil
 
 from telegram.ext import Updater, CommandHandler
 
-from yolobmsbot.batterypack import BatteryPack
+from yolobmsbot.batterypack import BatteryPack, BatteryCellValue
 from yolobmsbot.google import gsheets
 
 # bot settings
 SCRIPT_DIRECTORY = os.path.dirname(__file__)  # path to directory of python script running
 CLIENT_TOKEN_FILE = os.path.join(SCRIPT_DIRECTORY, "yolobmsbot", ".user_credentials", "telegram", "bot_token")  # path to token file
+UPDATE_INTERVAL_MINUTES = 30  # minutes between 2 consecutive values updates
 
 # chat settings
 KEYBOARD_CELLS = [[1, 2, 3], [4, 5, 6], [7, 8, 9], [10, 11, 12], [13, 14, 15], [16, 17, 18]]
 
 
-def get_bot_token():
-    """
-    :return: void
-        Authenticate bot
-    """
+class BatteryPackUpdater(object):
+    """ Fetches new values of battery pack and updates """
 
-    client_token = open(CLIENT_TOKEN_FILE, "r").read()
-    client_token = client_token.strip()
-    return client_token
+    def __init__(self, b_pack, update_interval):
+        """
+        :param b_pack: BatteryPack
+            Battery to update
+        :param update_interval: int
+            Number of minutes between two consecutive values updates
+        """
 
+        object.__init__(self)
 
-def get_keyboard(items, max_columns):
-    """
-    :param items: []
-        List of items to setup keyboard with
-    :param max_columns: int
-        Max columns per row
-    :return: [] of []
-        List of list: each row contains at most 3 items
-    """
+        self.battery_pack = b_pack  # battery
+        self.update_interval = update_interval
+        self.last_update = datetime.fromtimestamp(0)  # first january 1970
 
-    matrix = []
-    number_of_rows = ceil(len(items) / max_columns)
-    for r in range(number_of_rows):
-        row = []
-        for c in range(max_columns):  # loop through columns
-            try:
-                row.append(str(items[r * max_columns + c]))
-            except:
-                pass
-        matrix.append(row)
-    return matrix
+    def _is_update_needed(self):
+        """
+        :return: bool
+            True if need to update cells values because timer timeout
+        """
+
+        time_now = datetime.now()
+        minutes_since_last_update = time_now - self.last_update
+        minutes_since_last_update = minutes_since_last_update.total_seconds() / 60.0
+        return minutes_since_last_update >= self.update_interval
+
+    def update_values(self):
+        """
+        :return: void
+            Updates all values in battery pack
+        """
+
+        if self._is_update_needed():  # check timer timeout
+            self.last_update = datetime.now()  # update last time of update
+            for s in range(len(self.battery_pack.segments)):
+                last_values_of_cells_in_segment, last_update_time = gsheets.get_last_cells_values(s)  # get last values
+
+                for c in range(len(self.battery_pack.segments[s].cells)):
+                    new_voltage_value = last_values_of_cells_in_segment[c]
+                    new_temp_value = 0.0
+                    self.battery_pack.segments[s].cells[c].update_values(new_temp_value, new_voltage_value)  # update
 
 
 class YoloBmsBot(object):
@@ -83,6 +96,8 @@ class YoloBmsBot(object):
         self.dp.add_error_handler(self.error)  # log errors
         self.setup_commands()  # setup commands
         self.battery_pack = b_pack  # battery pack
+        self.values_updater = BatteryPackUpdater(self.battery_pack,
+                                                 UPDATE_INTERVAL_MINUTES)  # module to update cells values
 
     def run(self):
         """
@@ -129,7 +144,9 @@ class YoloBmsBot(object):
         """
 
         message_text = str(update.message.text)  # get text of user message
-        update.message.reply_text("Please, let me compute the query...")
+        update.message.reply_text("Please let me compute the query ... it may take a few moments")
+        self.values_updater.update_values()  # update values
+
         args = message_text.split(" ")
         if len(args) < 2:  # reply all segments
             for s in range(len(self.battery_pack.segments)):
@@ -150,11 +167,8 @@ class YoloBmsBot(object):
         """
 
         if self.battery_pack.is_cell_in_bounds(0, segment):
-            value, time = gsheets.get_last_segment_value(segment - 1)
-            time_date = time.split(" ")[0]
-            time_hours = time.split(" ")[1]
+            value, time_update = self.battery_pack.segments.get_average(BatteryCellValue.voltage)
             return value
-            # TODO update bms pack return self.battery_pack.segments.get_average(BatteryCellValue.voltage)
         else:
             raise ValueError("Invalid segment " + str(segment))
 
@@ -168,7 +182,11 @@ class YoloBmsBot(object):
 
         try:
             value = self.get_segment_value(segment)
-            return "Latest average value of segment " + str(segment) + " is " + "{0:.2f}".format(float(value)) + " mV"
+            time_update = self.values_updater.last_update
+            time_date, time_hours = str(time_update.date()), str(time_update.time())
+
+            return "Latest average value of segment " + str(segment) + " is " + "{0:.2f}".format(
+                float(value)) + " mV as of " + time_date + " at " + time_hours
         except Exception as e:
             return "Invalid segment " + str(segment)
 
@@ -181,7 +199,9 @@ class YoloBmsBot(object):
         """
 
         message_text = str(update.message.text)  # get text of user message
-        update.message.reply_text("Please, let me compute the query...")
+        update.message.reply_text("Please let me compute the query ... it may take a few moments")
+        self.values_updater.update_values()  # update values
+
         args = message_text.split(" ")
         if len(args) < 3:  # reply all cells
             for s in range(len(self.battery_pack.segments)):
@@ -205,11 +225,8 @@ class YoloBmsBot(object):
         """
 
         if self.battery_pack.is_cell_in_bounds():
-            value, time = gsheets.get_last_cell_value(cell - 1, segment - 1)
-            time_date = time.split(" ")[0]
-            time_hours = time.split(" ")[1]
+            value, time_update = self.battery_pack.segments[segment].cells[cell].get(BatteryCellValue.voltage)
             return value
-            # TODO update bms pack self.battery_pack.segments[segment].cells[cell].get(BatteryCellValue.voltage)
         else:
             raise ValueError("Invalid cell " + str(cell) + " in segment " + str(segment))
 
@@ -225,10 +242,48 @@ class YoloBmsBot(object):
 
         try:
             value = self.get_cell_value(cell, segment)
+            time_update = self.values_updater.last_update
+            time_date, time_hours = str(time_update.date()), str(time_update.time())
+
             return "Latest value of cell " + str(cell) + " in segment " + str(segment) + " is " + "{0:.2f}".format(
-                float(value)) + " mV"
+                float(value)) + " mV as of " + time_date + " at " + time_hours
         except Exception as e:
             return "Invalid cell " + str(cell) + " in segment " + str(segment)
+
+
+def get_bot_token():
+    """
+    :return: void
+        Authenticate bot
+    """
+
+    client_token = open(CLIENT_TOKEN_FILE, "r").read()
+    client_token = client_token.strip()
+    return client_token
+
+
+def get_keyboard(items, max_columns):
+    """
+    :param items: []
+        List of items to setup keyboard with
+    :param max_columns: int
+        Max columns per row
+    :return: [] of []
+        List of list: each row contains at most 3 items
+    """
+
+    matrix = []
+    number_of_rows = ceil(len(items) / max_columns)
+    for r in range(number_of_rows):
+        row = []
+        for c in range(max_columns):  # loop through columns
+            try:
+                row.append(str(items[r * max_columns + c]))
+            except:
+                pass
+        matrix.append(row)
+    return matrix
+
 
 if __name__ == '__main__':
     battery_pack = BatteryPack([18, 18, 18, 18, 18, 18])