+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.2.4 - 2017-04-09
+
+### Added
+- logging module
+- logging users
+
 ## 0.2.3 - 2017-04-09
 
 ### Added